Removed interfaces BaseDAO, ProfileDAO and TargetDAO
